LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/negative-layout-overflow-expected.txt


FAIL Anchor selection accounts for negative positioning. assert_equals: expected 350 but got 250
